Refactored Assets\Reader tests.
